Finish reorg
